refactor(hub): public-envelope to with-party/directory (spine keeps dynamic facade)
Moves kernel/meta/public-envelope/ into with-party/directory/public-envelope/
(E3 of the S5 family-doors effort). noydb.ts's setPublicEnvelope/
getPublicEnvelope facade methods now reach the service's runtime exports via
await import() instead of a static import, and the eagerly-resolved
publicEnvelopeSchema field is gone in favor of resolving it fresh (cheap,
module-cached) inside setPublicEnvelope. All other importers (root barrel,
with-pod/format.ts, with-pod/bundle.ts, kernel/types.ts, kernel/vault.ts,
the public-envelope test) just re-point to the new path.
