+++ packages/hub/__tests__/public-envelope.test.ts
@@ -16,7 +16,7 @@ import {
   isPublicEnvelope,
   PUBLIC_ENVELOPE_FIELDS,
   type PublicEnvelope,
-} from '../src/kernel/meta/public-envelope/index.js'
+} from '../src/with-party/directory/public-envelope/index.js'
 import { createNoydb } from '../src/kernel/noydb.js'
 
 function inlineMemory(): NoydbStore {
