+++ packages/hub/src/kernel/types.ts
@@ -54,7 +54,7 @@ import type { GuardStrategyHandleAny } from '../with-audit/guards/types.js'
 import type { DerivationStrategyHandle } from '../with-formula/derivations/types.js'
 import type { UnlockedKeyring } from '../with-party/team/keyring.js'
 import type { VaultPolicy } from './policy/types.js'
-import type { PublicEnvelopeSchema } from './meta/public-envelope/types.js'
+import type { PublicEnvelopeSchema } from '../with-party/directory/public-envelope/types.js'
 import type { MaterializedViewStrategyHandle } from '../with-formula/materialized-views/types.js'
 import type { OverlayedViewStrategyHandle } from '../with-formula/overlay-views/types.js'
 import type { SealingKeyProvider } from '../with-party/team/managed-passphrase.js'
