+++ packages/hub/src/with-party/directory/public-envelope/index.ts
