+++ packages/hub/src/with-pod/format.ts
@@ -46,7 +46,7 @@
  * bundle and verify its integrity — nothing about the contents.
  */
 
-import type { PublicEnvelope } from '../kernel/meta/public-envelope/types.js'
+import type { PublicEnvelope } from '../with-party/directory/public-envelope/types.js'
 
 /** Magic bytes 'NDB1' (ASCII), identifying a NOYDB bundle. */
 export const NOYDB_BUNDLE_MAGIC = new Uint8Array([0x4e, 0x44, 0x42, 0x31])
